Show broadcaster for feeds
Broadcaster is defined in schedule API, which has no direct reference to feeds defined in events API. For that reason it's not possible to know exact broadcaster if there are multiple broadcasters for same type of feed. Usually there's only one broadcaster though.

Adds also support for showing French feeds.